Documentos-Importantes/requirements.txt
opencv_python==4.9.0.80
opencv_python_headless==4.8.0.74
Pillow==10.1.0
Pillow==10.3.0
pypylon==3.0.0
pyserial==3.5
ultralytics==8.2.20
